renamed ResearchDetails to ResearchById
