+++ src/App.jsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom'
 import dashboard from '~/config/dashboard'
 
 import Login from '~/pages/Login'
-import ResearcherDetails from '~/pages/ReserachDetails'
+import ResearchById from '~/pages/ResearchById'
 
 const App = () => {
   return (
@@ -10,7 +10,7 @@ const App = () => {
       {dashboard.map(({ route, navigation }) => (
         <Route {...route} key={navigation.name} />
       ))}
-      <Route path="/research/:id" element={<ResearcherDetails />} />
+      <Route path="/research/:id" element={<ResearchById />} />
       <Route path="/login" element={<Login />} />
     </Routes>
   )
